+++ packages/x-date-pickers/src/PickersDay/PickersDay.spec.tsx
@@ -1,4 +1,4 @@
 import * as React from 'react';
 import { PickersDay } from '@mui/x-date-pickers/PickersDay';
 
-<PickersDay<Date> day={new Date()} outsideCurrentMonth onDaySelect={(date) => date?.getDay()} />;
+<PickersDay<Date> day={new Date()} outsideCurrentMonth />;
